SERVER-1 Moved SyslogMsg object to entities
